+++ PC2/Mode.cpp
@@ -21,7 +21,7 @@ void Mode::add_part_action(std::shared_ptr<Part>& part_ptr, std::string &action_
 	action_container.push_back(std::make_pair(part_ptr, action_name));
 }
 
-void Mode::add_sleep(std::string& part_name, int sleep_time)
+void Mode::add_sleep(std::string& part_name, float sleep_time)
 {
 	sleep_conatainer[part_name] = sleep_time;
 }
@@ -36,7 +36,7 @@ void Mode::do_mode()
 			LOG_ERROR << "write " << action.first->get_name() << " command failed!";
 		}
 		if (sleep_conatainer.find(action.first->get_name()) != sleep_conatainer.end()) {
-			int sleep_time = sleep_conatainer[action.first->get_name()];
+			float sleep_time = sleep_conatainer[action.first->get_name()];
 			LOG_ERROR << "sleep";
 			Sleep(sleep_time*1000);
 		}
